Add files via upload
A Stack is a linear data structure that follows the LIFO (Last-In-First-Out) principle. Stack has one end,  It contains only one pointer top pointer pointing to the topmost element of the stack. Whenever an element is added in the stack, it is added on the top of the stack, and the element can be deleted only from the stack. In other words, a stack can be defined as a container in which insertion and deletion can be done from the one end known as the top of the stack.
•push(): When we insert an element in a stack then the operation is known as a push. If the stack is full then the overflow condition occurs.
•pop(): When we delete an element from the stack, the operation is known as a pop. If the stack is empty means that no element exists in the stack, this state is known as an underflow state.